requirements.txt
antspyx
numpy
pandas
pydicom